+++ src/components/Icon/index.tsx
@@ -432,7 +432,7 @@ export { default as NotificationBell } from "./svg/NotificationBell"
 export { default as NotificationOffBell } from "./svg/NotificationOffBell"
 export { default as NoTraining } from "./svg/NoTraining"
 export { default as Number } from "./svg/Number"
-export { default as Object } from "./svg/Object"
+export { default as Object } from "./svg/ObjectIcon"
 export { default as On } from "./svg/On"
 export { default as OpenaiLogoBold } from "./svg/OpenaiLogoBold"
 export { default as OpenaiLogoBoldBoundingBox } from "./svg/OpenaiLogoBoldBoundingBox"
