+++ src/components/Icon/svg/ObjectIcon.tsx
